+++ src/ducks/Chart/Signals/helpers.js
@@ -2,10 +2,7 @@ import COLOR from '@santiment-network/ui/variables.scss'
 import { getTextWidth } from '@santiment-network/chart/utils'
 import { clearCtx } from '../utils'
 import { Metric } from '../../dataHub/metrics'
-import {
-  buildValueChangeSignal,
-  buildDAASignal
-} from '../../Signals/utils/utils'
+import { buildDAASignal } from '../../Signals/utils/utils'
 
 export const SIGNAL_BELOW = 'BELOW'
 export const SIGNAL_ABOVE = 'ABOVE'
@@ -137,17 +134,5 @@ export function makeSignalDrawable (
 export const checkPriceMetric = metric => metric === Metric.price_usd
 
 export const AlertBuilder = {
-  price_usd: (slug, value, type) => {
-    return buildValueChangeSignal(slug, value, type, Metric.price_usd)
-  },
-  daily_active_addresses: buildDAASignal,
-  social_volume_total: (slug, value, type) => {
-    return buildValueChangeSignal(slug, value, type, Metric.social_volume_total)
-  },
-  exchange_balance: (slug, value, type) => {
-    return buildValueChangeSignal(slug, value, type, Metric.exchange_balance)
-  },
-  transaction_volume: (slug, value, type) => {
-    return buildValueChangeSignal(slug, value, type, Metric.transaction_volume)
-  }
+  daily_active_addresses: buildDAASignal
 }
