+++ src/ducks/Studio/Alerts/suggestions/helpers.js
@@ -1,1 +1,41 @@
+import React from 'react'
+import { PRICE_CHANGE_TYPES } from '../../../Signals/utils/constants'
+import {
+  buildPercentUpDownSignal,
+  buildValueChangeSignal
+} from '../../../Signals/utils/utils'
+import Value from '../Value'
+
+export const SIGNAL_BELOW = 'BELOW'
+export const SIGNAL_ABOVE = 'ABOVE'
+export const VALUE_IFS = ['drops below', 'rises above']
+
 export const createSuggestion = (alert, render) => ({ alert, render })
+
+export const buildValueChangeSuggester = metric => {
+  const { formatter, label } = metric
+
+  return ({ slug, value, lastValue, ...rest }) => {
+    const isAbove = value > lastValue
+    const type = PRICE_CHANGE_TYPES[isAbove ? SIGNAL_ABOVE : SIGNAL_BELOW]
+
+    return createSuggestion(
+      buildValueChangeSignal(slug, value, type, metric),
+      <>
+        {label} {VALUE_IFS[+isAbove]} <Value>{formatter(value)}</Value>
+      </>
+    )
+  }
+}
+
+export const buildPercentUpSuggester = metric => {
+  const { label } = metric
+
+  return ({ slug }) =>
+    createSuggestion(
+      buildPercentUpDownSignal(slug, metric),
+      <>
+        {label} moves up or down by <Value>10%</Value>
+      </>
+    )
+}
